backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
redis
rq
soundfile
numpy
TTS
scipy
python-multipart 
pydub 
openai-whisper